Fix Roslynator diagnostics and integer overflow risks
Resolve 9 of 10 Roslynator diagnostics (CA1865, CA1861, CA1822, SYSLIB1045)
and fix manual code review issues: bare catches typed, null checks added to
EnumerateAllocationFrames, and Calls/AllocationCount widened from int to long
to eliminate int.MaxValue overflow guards.

+++ src/ProfileTool/Cpu/FunctionSample.cs
@@ -1,3 +1,3 @@
 namespace Asynkron.Profiler;
 
-public sealed record FunctionSample(string Name, double TimeMs, int Calls, int FrameIdx);
+public sealed record FunctionSample(string Name, double TimeMs, long Calls, int FrameIdx);
